Removed old agn folder and renamed agn_elasticc to agn. Removed unwanted line in setup
