+++ fink_science/agn/classifier.py
@@ -14,19 +14,26 @@
 # limitations under the License.
 
 import pickle
-import fink_science.agn.kernel as k
-import fink_science.agn.feature_extraction as fe
+import fink_science.agn_elasticc.kernel as k
+import fink_science.agn_elasticc.feature_extraction as fe
 import os
 from fink_science import __file__
 from fink_science.tester import spark_unit_tests
 import pandas as pd  # noqa: F401
 import numpy as np  # noqa: F401
+import fink_science.agn_elasticc.unit_examples as uex  # noqa: F401
 
 
-def load_classifier():
+def load_classifier(source):
     """
     load the random forest classifier trained to recognize the AGN
     on binary cases : AGNs vs non-AGNs  (pickle format).
+
+    Parameters
+    ----------
+    source: string
+        Origin of the data.
+        Currently accepts 'ZTF' or 'ELASTICC'.
 
     Returns
     -------
@@ -34,60 +41,69 @@ def load_classifier():
 
     Examples
     --------
-    >>> rf = load_classifier()
-    >>> rf.n_classes_
+    >>> rf_ELASTICC = load_classifier('ELASTICC')
+    >>> rf_ELASTICC.n_classes_
     2
-    >>> rf.n_features_
+    >>> rf_ELASTICC.n_features_
+    31
+    >>> rf_ZTF = load_classifier('ZTF')
+    >>> rf_ZTF.n_classes_
+    2
+    >>> rf_ZTF.n_features_
     12
     """
-    with open(k.CLASSIFIER, "rb") as f:
+
+    if source == 'ELASTICC':
+        model_path = k.CLASSIFIER_ELASTICC
+    elif source == 'ZTF':
+        model_path = k.CLASSIFIER_ZTF
+
+    with open(model_path, "rb") as f:
         clf = pickle.load(f)
 
     return clf
 
 
-def agn_classifier(data):
+def agn_classifier(data, source):
     """
-    call the agn_classifier
+    Call the agn_classifier
 
     Parameters
     ----------
     data : DataFrame
         alerts from fink with aggregated lightcurves
+    source: string
+        Origin of the data.
+        Currently accepts 'ZTF' or 'ELASTICC'.
 
     Returns
     -------
     np.array
         ordered probabilities of being an AGN
-        Return 0.0 if the minimum number of point per passband is not respected
+        Return 0 if the minimum number of point per passband is not respected
 
     Examples
     --------
-    >>> df = pd.read_parquet(ztf_alert_sample)
-    >>> proba = agn_classifier(df)
-    >>> len(proba)
-    2
-    >>> len(proba[proba != 0.0])
-    1
-    >>> len(proba[proba == 0.0])
-    1
+    >>> df = uex.raw_ztf_unit
+    >>> proba = agn_classifier(df, 'ZTF')
+    >>> proba[0] == 0
+    True
+    >>> proba[1] != 0
+    True
     """
 
-    clean = fe.clean_data(data)
-    converted = fe.convert_full_dataset(clean)
+    formated = fe.format_data(data, source)
 
-    transformed_1, transformed_2, valid = fe.transform_data(converted, k.MINIMUM_POINTS)
+    all_transformed, valid = fe.transform_data(formated, k.MINIMUM_POINTS, source)
 
-    all_empty = transformed_1.empty | transformed_2.empty
-    if all_empty:
+    if not valid.any():
         return np.zeros(len(data), dtype=np.float)
 
-    features_1 = fe.parametrise(transformed_1, 1)
-    features_2 = fe.parametrise(transformed_2, 2)
+    all_features = fe.parametrise(all_transformed, source)
 
-    features = fe.merge_features(features_1, features_2)
+    features = fe.merge_features(all_features, k.MINIMUM_POINTS, source)
 
-    clf = load_classifier()
+    clf = load_classifier(source)
 
     proba = fe.get_probabilities(clf, features, valid)
 
@@ -99,7 +115,7 @@ if __name__ == "__main__":
     globs = globals()
     path = os.path.dirname(__file__)
 
-    ztf_alert_sample = "{}/data/alerts/agn_example.parquet".format(path)
+    ztf_alert_sample = "{}/data/alerts/agn_elasticc_alerts.parquet".format(path)
     globs["ztf_alert_sample"] = ztf_alert_sample
 
     # Run the test suite
