+++ fink_science/agn/models.py
@@ -51,7 +51,7 @@ def bump(x, p1, p2, p3, p4):
     ----------
     x : np.array
         Array of mjd translated to 0
-    p1,p2,p3, p4 : floats
+    p1, p2, p3, p4 : floats
         Parameters of the function
 
     Returns
