+++ setup.py
@@ -45,8 +45,7 @@ setuptools.setup(
             'data/models/t2/*/*/variables*',
             'data/catalogs/*.parquet',
             'data/catalogs/*.fits',
-            'data/catalogs/*.csv',
-            'data/alerts/datatest/*.parquet'],
+            'data/catalogs/*.csv'],
     },
 
     classifiers=[
